Skip testing changes marked WIP or DNM
Check the first line of the commit message for either WIP (Work In
Progress) or DNM (Do Not Merge) and skip the rest of the job if it is
set.

This has the side effect of jenkins marking the change with a -1.
However this isn't necessarily bad as the change is otherwise untested.

Fixes: SCRD-5889